Finalise `File` cleanup
- Don't rewind or truncate streams in `File::copy()`
- Make `File::checkEof()` public
- Rename `File::getStream()` to `File::maybeOpen()` and make public
- Always `chmod()` after `mkdir()` in case umask modifies permissions
- Call `rewind($stream)` instead of `fseek($stream, 0)`
- Rename `File::getClosestExisting()` to `File::closestExisting()`
- Move `File::guessIndentation()` to `Indentation::from()`
